Summary of changes
In a nutshell, when the app is booted up for the first time, it will obtain its data from the web.
The app will then store all records in the local database via GreenDAO.

Changes made to the app

Repository  pattern implementation
I have decided to create a ComicRepository and have that injected into the GetComicsUseCase instead of injecting the ComicsService directly into the GetComicsUseCase.

This allows the GetComicsUseCase to use the ComicRepository, where the ComicRepository will decide where to obtain the data from.
Either this will come from the API via a web service request (and is subsequently stored in the local DB for future use)
or from the local database, if a saved entry exists.

Transaction during insertion
I have wrapped the entire Insert operation around a transaction. This is so all records are persisted or non at all. (TODO - better error handling for user facing client)

Updated existing entity objects
In order to accommodate GreenDao, I had to update existing entity objects and had to create OneToOne Bidirectional mappings, and OneToMany mappings.

Testing
Also added steho, in order to view the database whilst in debug mode.
Also added a test for the ComicRepository

TODO’s
My main todo is obviously, when these records are stored in the database, at some point the app would need to make fresh request to the API, in order to get the latest data from the web service
and then update the local database.